+++ packages/flutter_solidart/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.7.1
+
+### Changes from solidart
+
+- **FIX**: Expose `until` method for `Computed`.
+
 ## 2.7.0
 
 ### Changes from solidart
@@ -344,7 +350,7 @@ The core of the library has been rewritten in order to support automatic depende
     print('The counter changed from $previousValue to $value');
   }, fireImmediately: true);
   ```
-- **FEAT**: Add `firstWhere` method on `Signal`. It returns a future that completes when the condition evalutes to true and it returns the current signal value.
+- **FEAT**: Add `firstWhere` method on `Signal`. It returns a future that completes when the condition evaluates to true and it returns the current signal value.
   ```dart
   final count = createSignal(0);
   // wait until the count is greater than 5
@@ -572,7 +578,7 @@ The core of the library has been rewritten in order to support automatic depende
     print('The counter changed from $previousValue to $value');
   }, fireImmediately: true);
   ```
-- **FEAT**: Add `until` method on `Signal`. It returns a future that completes when the condition evalutes to true and it
+- **FEAT**: Add `until` method on `Signal`. It returns a future that completes when the condition evaluates to true and it
   returns the current signal value.
   ```dart
   final count = createSignal(0);